move clear_cuda_cache and related util fn to nvfuser module
