+++ benchmarks/python/test_dropout_rmsnorm_bwd.py
@@ -3,10 +3,9 @@
 # SPDX-License-Identifier: BSD-3-Clause
 import pytest
 from nvfuser import FusionDefinition, DataType
-from nvfuser.pytorch_utils import torch_dtype_to_nvfuser_dtype
+from nvfuser.pytorch_utils import torch_dtype_to_nvfuser_dtype, clear_cuda_cache
 from .core import (
     run_benchmark,
-    clear_cuda_cache,
     clear_dynamo_cache,
     unary_bwd_torch,
     compute_total_iobytes,
