+++ benchmarks/python/test_matmul.py
@@ -3,7 +3,8 @@
 # SPDX-License-Identifier: BSD-3-Clause
 import pytest
 from nvfuser import FusionDefinition
-from .core import run_benchmark, clear_cuda_cache
+from nvfuser.pytorch_utils import clear_cuda_cache
+from .core import run_benchmark
 import torch
 
 import csv
