+++ benchmarks/python/test_rope.py
@@ -3,7 +3,8 @@
 # SPDX-License-Identifier: BSD-3-Clause
 import pytest
 from nvfuser import FusionDefinition, DataType
-from .core import run_benchmark, clear_cuda_cache
+from nvfuser.pytorch_utils import clear_cuda_cache
+from .core import run_benchmark
 import torch
