+++ benchmarks/python/test_transformer.py
@@ -39,7 +39,8 @@ backward nvFusion executed many times.
 """
 
 from nvfuser import FusionDefinition, DataType
-from .core import run_benchmark, clear_cuda_cache
+from .core import run_benchmark
+from nvfuser.pytorch_utils import clear_cuda_cache
 import torch
