+++ tests/python/test_ops.py
@@ -8,7 +8,6 @@ import pytest
 import numpy as np
 from copy import deepcopy
 
-from benchmarks.python.core import clear_cuda_cache
 from opinfo_fusion_definitions import default_fd_fn, parse_inputs_fusion_definition
 from opinfo_framework import create_op_test, atexit_serde_create_op_test
 from opinfo_core import ReferenceType, OpInfo, SampleInput
@@ -17,6 +16,7 @@ from utils import ArgumentType, is_tensor, requiresJAX
 from typing import Callable
 
 from nvfuser import FusionCache, FusionDefinition
+from nvfuser.pytorch_utils import clear_cuda_cache
 
 from utils import check_captured_python_definition, debug_serde, basic_serde_check
